PookieBackend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
ecdsa==0.19.0
Faker==25.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.4
multipart==0.2.4
mysql-connector-python==8.3.0
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
wheel==0.43.0
geopy==2.2.0
